backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.26.0
pandas-datareader>=0.10.0

# Machine Learning & Statistics
scikit-learn>=1.3.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
supabase>=2.0.0

# AI/LLM Integration
google-generativeai>=0.8.0
google-genai>=1.0.0

# JWT
PyJWT>=2.8.0

# XBRL/EDINET Processing
requests>=2.31.0
lxml>=4.9.0

# Data Visualization (for backend chart generation)
plotly>=5.18.0

# Async Support
httpx>=0.25.0
aiohttp>=3.9.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
